Simplify code, move methods to more fitting places
(This commit includes a minor bug fix regarding visibility checks for
methods that are declared with `FamilyAndAssembly` accessibility,
which requires an additional check for `[InternalsVisibleTo]`.)
